Changed declaration of average to double

+++ src/basics1/CalculateAverage.java
@@ -18,7 +18,7 @@ public class CalculateAverage {
 		int number5 = new Scanner(System.in).nextInt();		
 		System.out.println("Please enter nunber6: ");
 		int number6 = new Scanner(System.in).nextInt();
-		int average =((number1 + number2 + number3 + number4 + number5 + number6)/6);
+		double average =((number1 + number2 + number3 + number4 + number5 + number6)/6);
 		System.out.println(average + " is the average of" + number1 + number2 + number3 + number4 + number5 + number6);
 	}
